Remove interactions with weather sensors (uses old data model, functionality moves to flexmeasures-openweathermap plugin): CLI function to add sensor and get forecasts
