+++ flexmeasures/utils/config_defaults.py
@@ -65,8 +65,6 @@ class Config(object):
     CORS_RESOURCES: Union[dict, list, str] = [r"/api/*"]
     CORS_SUPPORTS_CREDENTIALS: bool = True
 
-    OPENWEATHERMAP_API_KEY: Optional[str] = None
-
     MAPBOX_ACCESS_TOKEN: Optional[str] = None
 
     JSONIFY_PRETTYPRINT_REGULAR: bool = False
